added mypy support, mypy fixes, black-format files [CE-155]

+++ src/test/routers/test_link.py
@@ -7,11 +7,15 @@ from orcidlink.lib.config import config
 from orcidlink.main import app
 from orcidlink.model_types import LinkRecord
 from test.data.utils import load_data_file, load_data_json
-from test.mocks.mock_contexts import mock_auth_service, mock_orcid_oauth_service, no_stderr
+from test.mocks.mock_contexts import (
+    mock_auth_service,
+    mock_orcid_oauth_service,
+    no_stderr,
+)
 
 client = TestClient(app)
 
-config_yaml = load_data_file('config1.yaml')
+config_yaml = load_data_file("config1.yaml")
 
 
 @pytest.fixture
@@ -20,7 +24,7 @@ def fake_fs(fs):
     yield fs
 
 
-TEST_LINK = load_data_json('link1.json')
+TEST_LINK = load_data_json("link1.json")
 
 
 @contextlib.contextmanager
@@ -46,21 +50,19 @@ def create_link():
 # Tests
 #
 
+
 def test_get_link(fake_fs):
     with mock_services():
         create_link()
 
         client = TestClient(app)
-        response = client.get("/link",
-                              headers={"Authorization": "foo"})
+        response = client.get("/link", headers={"Authorization": "foo"})
         assert response.status_code == 200
 
-        response = client.get("/link",
-                              headers={"Authorization": "foox"})
+        response = client.get("/link", headers={"Authorization": "foox"})
         assert response.status_code == 401
 
-        response = client.get("/link",
-                              headers={"Authorization": "bar"})
+        response = client.get("/link", headers={"Authorization": "bar"})
         assert response.status_code == 404
 
 
@@ -69,18 +71,15 @@ def test_is_linked(fake_fs):
         create_link()
 
         client = TestClient(app)
-        response = client.get("/link/is_linked",
-                              headers={"Authorization": "foo"})
+        response = client.get("/link/is_linked", headers={"Authorization": "foo"})
         result = response.json()
         assert result is True
 
-        response = client.get("/link/is_linked",
-                              headers={"Authorization": "bar"})
+        response = client.get("/link/is_linked", headers={"Authorization": "bar"})
         result = response.json()
         assert result is False
 
-        response = client.get("/link/is_linked",
-                              headers={"Authorization": "baz"})
+        response = client.get("/link/is_linked", headers={"Authorization": "baz"})
         assert response.status_code == 401
 
 
@@ -92,10 +91,8 @@ def test_delete_link(fake_fs):
 
                 client = TestClient(app)
 
-                response = client.delete("/link",
-                                         headers={"Authorization": "foo"})
+                response = client.delete("/link", headers={"Authorization": "foo"})
                 assert response.status_code == 204
 
-                response = client.delete("/link",
-                                         headers={"Authorization": "bar"})
+                response = client.delete("/link", headers={"Authorization": "bar"})
                 assert response.status_code == 404
